with data

+++ riscv/sim.cc
@@ -459,7 +459,7 @@ void sim_t::sync_memory()
     flexicas::init_memory(dynamic_cast<mem_t*>(m.second)->get_map());
   }
   debug_mmu->set_init_memory();
-  for (int i = 0; i < procs.size(); i++){
+  for (uint64_t i = 0; i < procs.size(); i++){
     procs[i]->get_mmu()->set_init_memory();
   }
 }
